+++ lib/ts/recipe/emailverification/index.ts
@@ -61,9 +61,12 @@ export default class Wrapper {
         });
     }
 
-    static async sendEmail(input: TypeEmailVerificationEmailDeliveryInput & { userContext: any }) {
+    static async sendEmail(input: TypeEmailVerificationEmailDeliveryInput & { userContext?: any }) {
         let recipeInstance = Recipe.getInstanceOrThrowError();
-        return await recipeInstance.emailDelivery.ingredientInterfaceImpl.sendEmail(input);
+        return await recipeInstance.emailDelivery.ingredientInterfaceImpl.sendEmail({
+            userContext: {},
+            ...input,
+        });
     }
 }